Use continuations for faster Effect evaluation (#3)
This commit changes how readVar and writeVar work to provide a vastly faster translation step. The previous approach was to try and exhaustively enumerate all branches of all given Effects, and to collect the writes at the end. If we saw new writes, we would naively repeat the process - enumerating all Effects again, even if a very marginal variable had changed.

The change here is to instead use continuations to capture whenever a variable is read. On read, we capture the rest of the computation and "subscribe" it to the variable, such that whenever it is written to, the continuation is invoked with the new value. This means evaluation is compact - when a variable changes, only other Effects that read it will be re-evaluated, and even then only the new branches.