Prevail unchanged writes (#4)
This change means that if you do `readVar v >>= writeVar v`, this will be registered the same as just `readVar v` - that is, a prevailing condition for the effect. This allows one to internal use a variable to sequence an `Effect`, and then restore it to its original state at the end. This can be useful for temporarily swapping variables via some intermediary variable (that is always empty).