Correct detection of prevailing writes
In f12879e1e a new feature was added where readVar v >>= writeVar v would
be translated to just a prevailing condition, rather than an actual
write. However, the detection was wrong - rather than taking the
difference of writes - reads, we actually want to take the intersection
of writes and reads, and then filter this intersection for just
unchanged writes.