Prevent planner from locking up on large problems.
In some scenarios, the `downward` executable was producing a lot of
output which was not being read by the library. This would eventually
cause the downward process to hang, waiting for someone to consume its
output. This change drains both output handles before calling wait, and
thus preventing the deadlock.