Trivial solution for trivial goals.
The `[Test]` passed to the `solve` function is a conjuction of tests.
In a similar vein to `all`, an empty list should mean that the condition
was met. That is, all the conditions in the list were met, there just
weren't any conditions.

This comes up when you want to create a problem which is attempting to
find a plan between two `[a]`s. The `[]` to `[]` case will result in a
failure to solve due to no goals being set.